\docType{data}
\name{crocskull}
\alias{crocskull}
\title{3-D landmark data of crocodylian skulls}
\description{
Unaligned 3-D data of crocodylian skulls comprising 204 specimens and 78 landmarks.
}
\author{
Akinobu Watanabe
}
\references{
Watanabe, A., Slice, D. E. 2014. The utility of cranial ontogeny for phylogenetic inference: a case study in crocodylians using geometric morphometrics. Journal of Evolutionary Biology. 27: 1078–1092. doi: 10.1111/jeb.12382.
}
\keyword{datasets}