% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/package.R
\name{f}
\alias{f}
\title{Simple anonymous functions}
\usage{
f(expr, env = parent.frame())
}
\arguments{
\item{expr}{Expression that is the body of the function}

\item{env}{the environment the new function will have}
}
\description{
Function to create anonymous functions (lambdas) in a very concise manner.
Simply write the body of the function as you would normally.  If you want a
given variable to appear as one of the arguments to the function, put it in
\code{.()}.
}
\examples{
Reduce(f(.(x) + .(y)), 1:10)

compact1 <- f(FilterNegate(is.null(.(x))))

f(runif(n = rpois(1, 5), .(...)))
}

