\name{lambda-internal}
\title{Internal lambda objects}
\alias{.apply}
\alias{.cs}
\alias{.cumsum}
\alias{..cumsum}
\alias{.dim}
\alias{.dimnames}
\alias{.dup}
\alias{.get}
\alias{.l2a}
\alias{.na}
\alias{..na}
\alias{.rev}
\alias{..rev}
\alias{..seq}
\alias{.trv}
\alias{.w2L}
\alias{.warg}
\alias{.wD}
\alias{.win}
\alias{.wL}
\alias{.wNULL}
\alias{.wPCA}
\alias{.x}
\description{Internal lambda objects.}
\details{These are not to be called by the user.}
\keyword{internal}
