\docType{package}
\name{LaMBDA}
\alias{LaMBDA}
\alias{LaMBDA-package}
\title{LandMark-Based Data Assessment}
\description{
Assess the fidelity of landmark-based geometric morphometric data.
}
\author{
Akinobu Watanabe
}
\references{
Watanabe, A. How many landmarks are enough to characterize shape and size variation? ### UPDATE WHEN PUBLISHED ##
}
