% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda_fn.R
\name{lambda_fn}
\alias{lambda_fn}
\title{Calculate infinum of lambda}
\usage{
lambda_fn(lambda, prob, alpha = 0.05)
}
\arguments{
\item{df}{A dataframe with prob, theory, and participant.}

\item{threshold}{}
}
\description{
Calculate infinum of lambda
}
