% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda_operations.R
\name{lambda_get_event_source_mapping}
\alias{lambda_get_event_source_mapping}
\title{Returns details about an event source mapping}
\usage{
lambda_get_event_source_mapping(UUID)
}
\arguments{
\item{UUID}{[required] The identifier of the event source mapping.}
}
\description{
Returns details about an event source mapping. You can get the
identifier of a mapping from the output of ListEventSourceMappings.
}
\section{Request syntax}{
\preformatted{lambda$get_event_source_mapping(
  UUID = "string"
)
}
}

\examples{
# This operation retrieves a Lambda function's event source mapping
\donttest{lambda$get_event_source_mapping(
  UUID = "123489-xxxxx-kdla8d89d7"
)}

}
\keyword{internal}
