\docType{data}
\name{papioskull}
\alias{papioskull}
\title{3-D landmark data of baboon skulls}
\description{
Unaligned 3-D data of baboon skulls comprising 250 specimens and 231 landmarks.
}
\author{
Eric Delson
}
\references{
PRImate Morphomemetrics Online (PRIMO) database (http://primo.nycep.org).
}
\keyword{datasets}