\name{LAMBDA-package}
\alias{LAMBDA-package}
\alias{LAMBDA}
\docType{package}
\title{
\packageTitle{LAMBDA}
}
\description{
\packageDescription{LAMBDA}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{LAMBDA}
\packageIndices{LAMBDA}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{LAMBDA}

Maintainer: \packageMaintainer{LAMBDA}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
