\docType{package}
\name{LaMBDA}
\alias{LaMBDA}
\alias{LaMBDA-package}
\title{LandMark-Based Data Assessment}
\description{
Assess the fidelity of landmark-based geometric morphometric data.
}
\author{
Akinobu Watanabe
}
\references{
Watanabe, A. 2018. How many landmarks are enough to characterize shape and size variation? PLoS ONE 13(6): e0198341. https://doi.org/10.1371/journal.pone.0198341
}
