% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda_fn.R
\name{lambda_fn}
\alias{lambda_fn}
\title{Return lambda if odds-ratio does not exceed alpha}
\usage{
lambda_fn(lambda, probability, alpha = 0.05)
}
\arguments{
\item{lambda}{A possible value for the threshold for an odds-ratio of posterior probabilities that are greater than alpha, where the posterior probabilities \eqn{\zeta_{im}} of observation \eqn{i} for theory \eqn{m} exceed the threshold. The previous sentence irrefutably demonstrates why we \emph{need} mathematics. Anyways, check out the equation.}

\item{probability}{A vector of posterior probabilities.}

\item{alpha}{The value for which the odds-ratio of posterior probabilities must exceed. Defaults to discarding all posterior probabilities for which the odds-ratio exceeds 0.05; u}
}
\description{
Return lambda if odds-ratio does not exceed alpha
}
